% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_year.R
\name{calcAllYear}
\alias{calcAllYear}
\title{Summary metrics}
\usage{
calcAllYear(
  flow_matrix,
  max_nan_allowed_per_year = 100,
  do_plot = 1,
  user_q75_thresh = NULL,
  user_q90_thresh = NULL
)
}
\arguments{
\item{flow_matrix}{A 'by date' row data frame with 2 columns named (Date, Flow).}

\item{max_nan_allowed_per_year}{Maximum number of NA days per-year allowed. Recommend imputing NAs prior to running ffaus functions.}

\item{do_plot}{Should diagnostic plots be generated? 1 = yes, 0 = no.}

\item{user_q75_thresh}{A user defined 75th percentile threshold value.
In the same units of flow as in flow_matrix.}

\item{user_q90_thresh}{A user defined 90th percentile threshold value.
In the same units of flow as in flow_matrix.}
}
\value{
A list with elements
\itemize{
\item average_annual_flows - a one-row data frame with average flow for each
water year in each column.
\item standard_deviations
\item coefficient_variations
\item mags_global
\item mags
\item tims
\item freqs
\item durs
\item mags_long_evnt
\item tims_init_long_evnt
\item seas_init_long_evnt
\item durs_long_evnt
}
}
\description{
Computes by water year and over all year
summary statistics and exceedance metrics
}
