% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_season.R
\name{calcDrySeasonBaseflowDurationsMagnitude}
\alias{calcDrySeasonBaseflowDurationsMagnitude}
\title{Dry season base flow summary.}
\usage{
calcDrySeasonBaseflowDurationsMagnitude(
  flow_matrix,
  dry_season_start_dates,
  flow_initiation_flush_dates,
  return_to_wet_dates
)
}
\arguments{
\item{flow_matrix}{A 'by date' row data frame with 2 columns named (Date, Flow).}

\item{dry_season_start_dates}{The water year date for each water year in which the dry-season is
computed to have begun. Computed using calcStartOfdry_season().}

\item{flow_initiation_flush_dates}{The water year date for each water year in which the flow initiation
is computed to have begun. Corresponds typically to first substantial
peak in the hydrograph post dry-season. Computed using
calcFlowInitiationFlushTimingsDurations().}

\item{return_to_wet_dates}{The water year date for each water year in which the flow is beginning to
return to its wetter period. Computed using returnToWetDate().}
}
\value{
A list with elements
\itemize{
\item dry_season_90_magnitudes
\item dry_season_50_magnitudes
\item dry_season_10_magnitudes
\item dry_season_flush_durations
\item dry_season_wet_durations
\item dry_season_no_flow_counts
}
}
\description{
Computed dry-season baseflow duration and magnitude
metrics
}
