% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_initiation.R
\name{calcFlowInitiationFlushTimingsDurations}
\alias{calcFlowInitiationFlushTimingsDurations}
\title{Initiation flow timing and duration.}
\usage{
calcFlowInitiationFlushTimingsDurations(
  flow_matrix,
  class_number = 3,
  max_nan_allowed_per_year = 100,
  dry_season_start_dates,
  min_flow_rate = 1,
  sigma = 2,
  broad_sigma = 15,
  wet_season_sigma = 12,
  peak_sensitivity = 0.005,
  peak_sensitivity_wet = 0.005,
  max_flush_duration = 40,
  wet_threshold_perc = 0.2,
  peak_detect_perc = 0.3,
  flush_threshold_perc = 0.3,
  min_flush_threshold = 1,
  last_date_cutoff_flush = 152,
  slope_sensitivity = 300,
  do_plot = 1
)
}
\arguments{
\item{flow_matrix}{a 'by date' row data frame with 2 columns named (Date, Flow).}

\item{class_number}{a remnant from Python implementation. Don't use.}

\item{max_nan_allowed_per_year}{maximum number of NA days per-year allowed. Recommend imputing NAs prior to running ffaus functions.}

\item{dry_season_start_dates}{the water year date for each water year in which the dry-season is
computed to have begun. Computed using calcStartOfdry_season().}

\item{min_flow_rate}{Don't calculate flow metrics if max flow is below this value.}

\item{sigma}{Standard deviation for Gaussian kernel smoothing of within year hydrograph used for filtering.}

\item{broad_sigma}{Larger standard deviation filter to provide heavy smooth of within year hydrograph to find wet season peak}

\item{wet_season_sigma}{Medium standard deviation filter to provide intermediate smooth of within year hydrograph to find wet season initiation peak}

\item{peak_sensitivity}{Sensitivity of peak detection algorithm.}

\item{peak_sensitivity_wet}{Sensitivity of peak detection algorithm for detecting peaks of wet-season smoothed data.}

\item{max_flush_duration}{Maximum duration in days from start to end, for initiation/flush episode.}

\item{wet_threshold_perc}{Return to wet season flow must be 'this' percentage of that year's max flow.}

\item{peak_detect_perc}{The peak percentage of max flow identified to search before for wet season initiation.}

\item{flush_threshold_perc}{Flow at half duration of peak must be 'this' proportion of the maximum.}

\item{min_flush_threshold}{Minimum allowable magnitude threshold for initiation flow}

\item{last_date_cutoff_flush}{Latest accepted date for winter spring flush, in Julian Date counting from Jul 1st = 1. (i.e., 2017-11-30)}

\item{slope_sensitivity}{Sets sensitivity of slope threshold for wet season start time.}

\item{do_plot}{Should diagnostic plots be generated? 1 = yes, 0 = no.}
}
\value{
A list with data frame elements of size 1 by no. water years
\itemize{
\item start_dates - flow initiation start dates.
\item mags - the peak magnitude of flow associated with the flow initiation start dates.
\item wet_dates - the date in which the wet season is deemed to have started for that water year.
\item durations - the duration of the flow initiation peak, which is typically the no. of days to from the left minimum of the peak to the peak.
}
}
\description{
Summarises the first return of water to the hydrograph following the dry season
for each water year.
}
