% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highflow.R
\name{calcHighflowAnnual}
\alias{calcHighflowAnnual}
\title{Water year high flow exccedance and shape}
\usage{
calcHighflowAnnual(
  flow_matrix,
  exceedance_percent = c(98, 95, 90, 80),
  max_nan_allowed_per_year = 100,
  do_plot = 1
)
}
\arguments{
\item{flow_matrix}{a 'by date' row data frame with 2 columns named (Date, Flow).}

\item{exceedance_percent}{vector of length 1 by 4 containing the exceedance percentages for calculating thresholds.}

\item{max_nan_allowed_per_year}{maximum number of NA days per-year allowed. Recommend imputing NAs prior to running ffaus functions.}

\item{do_plot}{should diagnostic plots be generated? 1 = yes, 0 = no.}
}
\value{
A list with two data frame elements of size 1 by no. water years
\itemize{
\item timing - the median over the exceedance event start dates.
\item duration - Sum over the days in which flow exceeds the Q∗ threshold.
\item freq - the number of events that exceed the Q∗ threshold. An event has to be at lest one
day and a gap of three days to be a unique event.
\item magnitude - for all events that exceed the Q∗ threshold take the median over the maximum values for each event.
\item magnitude_exceed - the exceedance thresholds for each of the exceedance_percent vector percentiles.
\item peak_magnitude - the peak magnitude for that water year.
\item peak_magnitude_tim - the water year day of peak magnitude.
}
}
\description{
Summarises the initiation and magnitude of the flow recession,
which corresponds to the final peak and decay of the hydrograph
to the dry season. For the following metrics we compute the 80th,
90th, 95th, and 98th percentiles of the flow data (Q*) for the whole period of
flow. These are then used as exceedance values for the computation
of the following metrics. For each magnitude threshold the flow
events above these thresholds are summarised.
}
