% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_recession.R
\name{calcRecessionDuration}
\alias{calcRecessionDuration}
\title{Flow recession duration.}
\usage{
calcRecessionDuration(recession_timings, dry_season_start_dates)
}
\arguments{
\item{recession_timings}{vector of flow recession start dates for each water year.}

\item{dry_season_start_dates}{the water year date for each water year in which the dry-season is
computed to have begun. Computed using calcStartOfdry_season().}
}
\value{
A list with two data frame elements of size 1 by no. water years
\itemize{
\item duration - the length of the flow recession in days for each water year.
}
}
\description{
Summarises the number of water days for the flow
to recede from the final peak to the start of the dry season.
}
