% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_recession.R
\name{calcRecessionRoc}
\alias{calcRecessionRoc}
\title{Flow recession duration.}
\usage{
calcRecessionRoc(flow_matrix, recession_timings, dry_season_start_dates)
}
\arguments{
\item{flow_matrix}{a 'by date' row data frame with 2 columns named (Date, Flow).}

\item{recession_timings}{vector of flow recession start dates for each water year.}

\item{dry_season_start_dates}{the water year date for each water year in which the dry-season is
computed to have begun. Computed using calcStartOfdry_season().}
}
\value{
A list with two data frame elements of size 1 by no. water years
\itemize{
\item rocs_only_neg - the median over flow recession days of the recession slope as a percentage change from previous day.
}
}
\description{
Summarises the number of water days for the flow
to recede from the final peak to the start of the dry season.
}
