% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_recession.R
\name{calcRecessionTimingMagnitude}
\alias{calcRecessionTimingMagnitude}
\title{Flow recession timing and magnitude.}
\usage{
calcRecessionTimingMagnitude(
  flow_matrix,
  dry_season_start_dates,
  class_number = 3,
  max_nan_allowed_per_year = 100,
  max_peak_flow_date = 300,
  search_window_left = 20,
  search_window_right = 50,
  peak_sensitivity = 0.1,
  peak_filter_percentage = 0.5,
  min_max_flow_rate = 0.1,
  window_sigma = 10,
  fit_sigma = 1.3,
  sensitivity = 0.2,
  min_percentage_of_max_flow = 0.5,
  timing_cutoff = 80,
  min_flow_rate = 1,
  do_plot = 1
)
}
\arguments{
\item{flow_matrix}{a 'by date' row data frame with 2 columns named (Date, Flow).}

\item{dry_season_start_dates}{the water year date for each water year in which the dry-season is
computed to have begun. Computed using calcStartOfdry_season().}

\item{class_number}{a remnant from Python implementation. Don't use.}

\item{max_nan_allowed_per_year}{maximum number of NA days per-year allowed. Recommend imputing NAs prior to running ffaus functions.}

\item{max_peak_flow_date}{maximum search date for the final peak before flow recedes.}

\item{search_window_left}{the number of search window days to the left of the peak before flow recedes.}

\item{search_window_right}{the number of search window days to the right of the peak before flow recedes.}

\item{peak_sensitivity}{the sensitivity used for peak detection algorithm. Smaller values leads to more peaks detected.}

\item{peak_filter_percentage}{the proportion of the flow range (Qmax - Qmin) that the final peak must exceed to be considered the final peak.}

\item{min_max_flow_rate}{If filtered max flow is below this, automatically set recession timing to date of max flow.}

\item{window_sigma}{heavy standard deviation for Gaussian kernel smoothing to identify major peaks in entire water year.}

\item{fit_sigma}{smaller standard deviation  to identify small peaks in windowed data.}

\item{sensitivity}{controls the threshold as a proportion of maximum flow for first order derivative threshold. 0.1 - 10, 10 being the most sensitive.}

\item{min_percentage_of_max_flow}{the detected date's flow in the window around the peak has be certain percentage of the max flow in the window.}

\item{timing_cutoff}{earliest accepted date at which recession can begin.}

\item{min_flow_rate}{minimum allowable magnitude threshold for fall flush flow}

\item{do_plot}{Should diagnostic plots be generated? 1 = yes, 0 = no.}
}
\value{
A list with two data frame elements of size 1 by no. water years
\itemize{
\item timings - the start date of the flow recession.
\item magnitudes - the magnitude of flow at recession start date.
}
}
\description{
Summarises the initiation and magnitude of the flow recession,
which corresponds to the final peak and decay of the hydrograph
to the dry season.
}
