% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_season.R
\name{calcStartOfdry_season}
\alias{calcStartOfdry_season}
\title{Start of dry season}
\usage{
calcStartOfdry_season(
  flow_matrix,
  class_number,
  max_nan_allowed_per_year = 100,
  sigma = 5,
  sensitivity = 800,
  peak_sensitivity = 0.2,
  max_peak_flow_date = 250,
  min_dry_season_flow_percent = 0.125,
  min_flow_rate = 1,
  do_plot = 1
)
}
\arguments{
\item{flow_matrix}{A 'by date' row data frame with 2 columns named (Date, Flow).}

\item{class_number}{A remnant from Python implementation. Don't use.}

\item{max_nan_allowed_per_year}{Maximum number of NA days per-year allowed.}

\item{sigma}{Scalar to set amount of smoothing.}

\item{sensitivity}{Increased sensitivity returns smaller threshold for derivative.}

\item{peak_sensitivity}{Identifies last major peak after which to search for start date.}

\item{max_peak_flow_date}{Max search date for the final peak flow date of each water year.}

\item{min_dry_season_flow_percent}{Percentile of flow taken to be the threshold for dry-season start. Percentile is computed over
flow values from the final peak to the end of the water-year.}

\item{min_flow_rate}{Don't calculate flow metrics if max flow is below this value.}

\item{do_plot}{Should diagnostic plots be generated? 1 = yes, 0 = no.}
}
\value{
A data frame of size 1 by no. water years with the dry-season start date for each
water year.
}
\description{
Computes the start of the dry season for each water year.
}
