% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highflow.R
\name{calcWhippleHighFlow}
\alias{calcWhippleHighFlow}
\title{Highflow/flood shape and frequency of peaks.}
\usage{
calcWhippleHighFlow(flow_matrix, whipple_exceed_thresh = 0.95)
}
\arguments{
\item{flow_matrix}{a 'by date' row data frame with 2 columns named (Date, Flow).}

\item{whipple_exceed_thresh}{percentile threshold to compute flow value from flow data over all years.}
}
\value{
A data frame of size 5 by no. water years named whipple_shape. Each row corresponds to
metric
\itemize{
\item No_Events  -  Number of events that exceed the threshold defined by whipple_exceed_thresh
\item No_Peaks_Avg    - The average number of peaks within exceedance events.
\item No_Peaks_Max    - The maximum number of inter-event peaks.
\item Frac_V_Cent_Avg - Over all events the mean value of the number of days to > 0.5 of total flood event volume divided by the total number event days.
\item Frac_V_Cent_Max - For the event that contains the maximum flow for the water year, the value of the number of days to > 0.5 of total flood event volume divided by the total number event days.
}
}
\description{
We further summarise peak flow shape and peak frequency hydrometrics functions from
Allison Whipple’s hydrospatial R package. The code rests on passing a user defined percentile
value in which to summarise the peaks of each year’s hydrograph using the whipple_exceed_thresh
parameter, with the default being 0.95.
}
