% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFFmetrics.R
\name{getFFMetrics}
\alias{getFFMetrics}
\title{Global hydrological metrics calculation and exceedance summary
function.}
\usage{
getFFMetrics(
  flow_matrix,
  class_number = 3,
  do_plot = 0,
  just_all_year = NULL,
  user_q75_thresh = NULL,
  user_q90_thresh = NULL,
  plot_out_base = NULL,
  dry_max_peak_flow_date = 250,
  dry_sensitivity = 800,
  min_dry_season_flow_percent = 0.125,
  init_broad_sigma = 15,
  last_date_cutoff_initiation = 152,
  wet_peak_detect_perc = 0.3,
  whipple_exceed_thresh = 0.95,
  recess_init_timing_cutoff = 80,
  recess_peak_filter_percentage = 0.5
)
}
\arguments{
\item{flow_matrix}{a 'by date' row data frame with 2 columns named (Date, Flow).}

\item{class_number}{a remnant from Python implementation. Don't use.}

\item{do_plot}{should diagnostic plots be generated? 1 = yes, 0 = no.}

\item{just_all_year}{if = 1 then only returns all year summaries with
exceedance metrics i.e., no functional flow hydrometrics.}

\item{user_q75_thresh}{pass a user defined 75th percentile threshold for use in calculating exceedance metrics.}

\item{user_q90_thresh}{pass a user defined 90th percentile threshold for use in calculating exceedance metrics.}

\item{plot_out_base}{file path for diagnostics plots to be written to.}

\item{dry_max_peak_flow_date}{the last date in which to begin looking for final peak before dry season. A key parameter to adjust if
dry season start dates are not being detected correctly.}

\item{dry_sensitivity}{controls the sensitivity of the slope threshold for the dry season start date cutoff. Smaller values allow to find curvier (:)) start dates.}

\item{min_dry_season_flow_percent}{Percentile of flow taken to be the threshold for dry-season start. Percentile is computed over
flow values from the final peak to the end of the water-year.}

\item{init_broad_sigma}{the standard deviation of the Gaussian kernel used to smooth intra-annual flow for detection of the flow initiation. A key parameter to adjust if
flow initiation start dates are not being detected correctly. Default is 15 so make small adjustments first either side.}

\item{last_date_cutoff_initiation}{the last date that flow initiation can begin. Can be later in water year for flashy water systems.}

\item{wet_peak_detect_perc}{the proportion of the maximum yearly (water-year) flow that the return to wet peak must be to be considered the peak that initiates the wet period.}

\item{whipple_exceed_thresh}{the exceedance percentile threshold for large flood flows.}

\item{recess_init_timing_cutoff}{earliest accepted date of water year in which the last peak before the recession can occur.}

\item{recess_peak_filter_percentage}{the proportion of the flow range (Qmax - Qmin) that the final peak must exceed
to be considered the final peak for detection of recession start. A key parameter to adjust if recession start dates are not as expected.}
}
\value{
A list with one data frame (at this stage) named FFmetrics
\itemize{
\item The data frame FFmetrics includes the exceedance metrics and the functional
flow metrics in the rows and the columns are the summary value
for each water year.
}
}
\description{
Primary ffaus function that takes typically daily flow data
as a 'by date' row data frame with 2 columns named (Date, Flow).
The getFFMetrics function calls each of the functional flow
component functions and returns a summary data frame of the
containing the hydrometrics by row and water year summarised
in columns.
}
