% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{makeFlowMatrix}
\alias{makeFlowMatrix}
\title{Flow to ffaus flow format.}
\usage{
makeFlowMatrix(flow_matrix, water_year_st, impute = NULL)
}
\arguments{
\item{flow_matrix}{a 'by date' row data frame with 2 columns named (Date, Flow).
The colnames need to be Date and Flow.}

\item{water_year_st}{a character date in format \%m-\%d that specifies the month and day start of the water year.
For example "07-01" is the first of the July water year start date.}

\item{impute}{would you like to impute missing values and dates? Only recommend if less than a handful per year.}
}
\value{
A list with two data frame elements
\itemize{
\item flow_deep - Date (starting at first water year day) by 8 column object that contains in each row\preformatted{        * Date - in \%Y-\%m-\%d format.
        * Flow - in units of data provided.
        * Season - summer (sum), autumn (aut), winter (win), or spring (spr).
        * Month (Julian) - Month as an interger 1-12 (Jan-Dec).
        * Year - Integer for year.
        * Year_Day - Julian day of year
        * Water_Year - Integer for water year.
        * Water_Year_Day - Integer for water year day relative to water_year_st.
}
}
}
\description{
Function that takes flow data (Date, Flow) and processes to
ffaus flow format.
}
